requirements.txt
# Multi-Language TTS Training System Requirements
# Core dependencies for extensible TTS training

# Core Python packages
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Audio processing
librosa>=0.9.0
soundfile>=0.10.0
audioread>=2.1.9

# Configuration and data handling
PyYAML>=6.0

# Text processing
regex>=2021.8.3

# Machine learning (install separately for GPU support)
# torch>=2.0.0
# torchaudio>=2.0.0

# TTS Models (install as needed)
# TTS>=0.22.0  # Coqui TTS for VITS
# edge-tts>=6.1.0  # Microsoft Edge TTS

# Plugin dependencies (install as needed)
# transformers>=4.30.0  # For Orpheus, StyleTTS
# accelerate>=0.20.0    # For GPU acceleration
# datasets>=2.12.0      # For dataset handling

# Evaluation and analysis (optional)
scikit-learn>=1.0.0
matplotlib>=3.4.0
tqdm>=4.62.0

# Logging
colorlog>=6.4.0

# Utilities
click>=8.0.0

# Security updates
certifi>=2021.5.25
urllib3>=1.26.6

# Installation Instructions:
#
# 1. Basic installation:
#    pip install -r requirements.txt
#
# 2. For GPU training (recommended):
#    pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu121
#    pip install TTS
#
# 3. For specific models:
#    pip install transformers accelerate  # For Orpheus, StyleTTS
#    pip install edge-tts                 # For Edge TTS
#
# 4. For development:
#    pip install jupyter notebook tensorboard
